+++ pkg/sylog/sylog_dummy_test.go
@@ -25,7 +25,7 @@ func TestGetLevel(t *testing.T) {
 	defer test.ResetPrivilege(t)
 
 	l := GetLevel()
-	if l != -1 {
+	if l != 1 {
 		t.Fatalf("%d was returned instead of -1", l)
 	}
 }